+++ src/nftpServer.c
@@ -109,7 +109,7 @@ main (int argc, char** argv){
        
   close_srvctcp(sk);
 
-  printf("Total bytes sent %d\n", total_bytes_sent);
+  printf("Total bytes sent %d\n", (int) total_bytes_sent);
      
   fclose(snd_file);
